+++ kas-wgpu/src/theme/mod.rs
@@ -1,0 +1,13 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License in the LICENSE-APACHE file or at:
+//     https://www.apache.org/licenses/LICENSE-2.0
+
+//! Themes
+
+mod sample_theme;
+mod size_handle;
+
+pub(crate) use size_handle::{Dimensions, DimensionsParams, SizeHandle};
+
+pub use sample_theme::SampleTheme;
